decimal values in NewPodRab and NewVidRab

+++ Manufactory/NewPodRab.cs
@@ -26,7 +26,7 @@ namespace Manufactory
             this.textBox8.ReadOnly = true;
             this.textBox8.Text = 0.ToString();
             this.textBox1.TextChanged += (sender, e) => showCost(sender, e, textBox1, textBox5, 17);
-            this.textBox2.TextChanged += (sender, e) => showCost(sender, e, textBox2, textBox6, 15.5);
+            this.textBox2.TextChanged += (sender, e) => showCost(sender, e, textBox2, textBox6, 15.5m);
             this.textBox3.TextChanged += (sender, e) => showCost(sender, e, textBox3, textBox7, 12);
             this.textBox4.TextChanged += (sender, e) => showCost(sender, e, textBox4, textBox8, 25);
         }
@@ -61,12 +61,12 @@ namespace Manufactory
         {
             throw new NotImplementedException();
         }
-        protected void showCost(object sender, EventArgs e, Control inputControl, Control outputControl, double price)//TODO: Поменять на decimal
+        protected void showCost(object sender, EventArgs e, Control inputControl, Control outputControl, decimal price)//TODO: Поменять на decimal
         {
 
             try
             {
-                outputControl.Text = (Convert.ToDouble(inputControl.Text) * price).ToString();
+                outputControl.Text = (Convert.ToDecimal(inputControl.Text) * price).ToString();
             }
             catch (FormatException)
             {
